Update import paths for components and types
