--- components/top-components/home.tsx
@@ -1,12 +1,0 @@
-// Import necessary components and libraries
-import CocktailsHome from "./cocktails.home";
-
-const HomeSection = () => {
-	return (
-		<section className="flex flex-col w-full">
-			<CocktailsHome />
-		</section>
-	);
-};
-
-export default HomeSection;
