--- components/top-components/search/search.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import CardCocktail from '@/components/common/card'
-import { searchResultsAtom } from '@/components/atoms'
-import { useAtom } from 'jotai'
-
-export default function SearchPage() {
-  return (
-    <>
-    
-    </>
-  )
-}
