+++ src/components/auth/authenticationSection.tsx
